FuzzSys-CQA/requirements.txt
numpy==1.24.4
pandas==2.2.2
PyYAML==6.0
scikit-learn==1.1.3
scipy==1.13.1
torch==2.2.2
torchmetrics==1.4.0.post0
